move gete_graph_name to gs.utils
