+++ python/graphstorm/utils.py
@@ -30,6 +30,22 @@ import numpy as np
 TORCH_MAJOR_VER = int(th.__version__.split('.', maxsplit=1)[0])
 USE_WHOLEGRAPH = False
 
+def get_graph_name(part_cnfig):
+    """ Get graph name from graph partition config file
+
+    Parameter
+    ---------
+    part_config: str
+        Path to graph partition config file
+
+    Return
+    ------
+        graph_name
+    """
+    with open(part_cnfig, "r", encoding='utf-8') as f:
+        config = json.load(f)
+    return config["graph_name"]
+
 def setup_device(local_rank):
     r"""Setup computation device.
